// proto/osmformat.ts
export interface IHeaderBlock {
    bbox?: IHeaderBBox;
    required_features?: string[];
    optional_features?: string[];
    writingprogram?: string;
    source?: string;
    osmosis_replication_timestamp?: number;
    osmosis_replication_sequence_number?: number;
    osmosis_replication_base_url?: string;
}

export interface IHeaderBBox {
    left: number;
    right: number;
    top: number;
    bottom: number;
}

export interface IPrimitiveBlock {
    stringtable: IStringTable;
    primitivegroup?: IPrimitiveGroup[];
    granularity?: number;
    lat_offset?: number;
    lon_offset?: number;
    date_granularity?: number;
}

export interface IPrimitiveGroup {
    nodes?: INode[];
    dense?: IDenseNodes;
    ways?: IWay[];
    relations?: IRelation[];
    changesets?: IChangeSet[];
}

export interface IStringTable {
    s?: Uint8Array[];
}
// ...
export const HeaderBlock = {
    read(pbf: Pbf, end?: number): IHeaderBlock {
        return pbf.readFields(HeaderBlock._readField, {bbox: null, required_features: [], optional_features: [], writingprogram: "", source: "", osmosis_replication_timestamp: 0, osmosis_replication_sequence_number: 0, osmosis_replication_base_url: ""}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.bbox = HeaderBBox.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 4) obj.required_features.push(pbf.readString());
        else if (tag === 5) obj.optional_features.push(pbf.readString());
        else if (tag === 16) obj.writingprogram = pbf.readString();
        else if (tag === 17) obj.source = pbf.readString();
        else if (tag === 32) obj.osmosis_replication_timestamp = pbf.readVarint(true);
        else if (tag === 33) obj.osmosis_replication_sequence_number = pbf.readVarint(true);
        else if (tag === 34) obj.osmosis_replication_base_url = pbf.readString();
    },
    write(obj: IHeaderBlock, pbf: Pbf): void {
        if (obj.bbox) pbf.writeMessage(1, HeaderBBox.write, obj.bbox);
        if (obj.required_features) for (var i = 0; i < obj.required_features.length; i++) pbf.writeStringField(4, obj.required_features[i]);
        if (obj.optional_features) for (i = 0; i < obj.optional_features.length; i++) pbf.writeStringField(5, obj.optional_features[i]);
        if (obj.writingprogram) pbf.writeStringField(16, obj.writingprogram);
        if (obj.source) pbf.writeStringField(17, obj.source);
        if (obj.osmosis_replication_timestamp) pbf.writeVarintField(32, obj.osmosis_replication_timestamp);
        if (obj.osmosis_replication_sequence_number) pbf.writeVarintField(33, obj.osmosis_replication_sequence_number);
        if (obj.osmosis_replication_base_url) pbf.writeStringField(34, obj.osmosis_replication_base_url);
    }
};

export const HeaderBBox = {
    read(pbf: Pbf, end?: number): IHeaderBBox {
        return pbf.readFields(HeaderBBox._readField, {left: 0, right: 0, top: 0, bottom: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.left = pbf.readSVarint();
        else if (tag === 2) obj.right = pbf.readSVarint();
        else if (tag === 3) obj.top = pbf.readSVarint();
        else if (tag === 4) obj.bottom = pbf.readSVarint();
    },
    write(obj: IHeaderBBox, pbf: Pbf): void {
        if (obj.left) pbf.writeSVarintField(1, obj.left);
        if (obj.right) pbf.writeSVarintField(2, obj.right);
        if (obj.top) pbf.writeSVarintField(3, obj.top);
        if (obj.bottom) pbf.writeSVarintField(4, obj.bottom);
    }
};

export const PrimitiveBlock = {
    read(pbf: Pbf, end?: number): IPrimitiveBlock {
        return pbf.readFields(PrimitiveBlock._readField, {stringtable: null, primitivegroup: [], granularity: 100, lat_offset: 0, lon_offset: 0, date_granularity: 1000}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.stringtable = StringTable.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 2) obj.primitivegroup.push(PrimitiveGroup.read(pbf, pbf.readVarint() + pbf.pos));
        else if (tag === 17) obj.granularity = pbf.readVarint(true);
        else if (tag === 19) obj.lat_offset = pbf.readVarint(true);
        else if (tag === 20) obj.lon_offset = pbf.readVarint(true);
        else if (tag === 18) obj.date_granularity = pbf.readVarint(true);
    },
    write(obj: IPrimitiveBlock, pbf: Pbf): void {
        if (obj.stringtable) pbf.writeMessage(1, StringTable.write, obj.stringtable);
        if (obj.primitivegroup) for (var i = 0; i < obj.primitivegroup.length; i++) pbf.writeMessage(2, PrimitiveGroup.write, obj.primitivegroup[i]);
        if (obj.granularity != undefined && obj.granularity !== 100) pbf.writeVarintField(17, obj.granularity);
        if (obj.lat_offset) pbf.writeVarintField(19, obj.lat_offset);
        if (obj.lon_offset) pbf.writeVarintField(20, obj.lon_offset);
        if (obj.date_granularity != undefined && obj.date_granularity !== 1000) pbf.writeVarintField(18, obj.date_granularity);
    }
};

export const PrimitiveGroup = {
    read(pbf: Pbf, end?: number): IPrimitiveGroup {
        return pbf.readFields(PrimitiveGroup._readField, {nodes: [], dense: null, ways: [], relations: [], changesets: []}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.nodes.push(Node.read(pbf, pbf.readVarint() + pbf.pos));
        else if (tag === 2) obj.dense = DenseNodes.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 3) obj.ways.push(Way.read(pbf, pbf.readVarint() + pbf.pos));
        else if (tag === 4) obj.relations.push(Relation.read(pbf, pbf.readVarint() + pbf.pos));
        else if (tag === 5) obj.changesets.push(ChangeSet.read(pbf, pbf.readVarint() + pbf.pos));
    },
    write(obj: IPrimitiveGroup, pbf: Pbf): void {
        if (obj.nodes) for (var i = 0; i < obj.nodes.length; i++) pbf.writeMessage(1, Node.write, obj.nodes[i]);
        if (obj.dense) pbf.writeMessage(2, DenseNodes.write, obj.dense);
        if (obj.ways) for (i = 0; i < obj.ways.length; i++) pbf.writeMessage(3, Way.write, obj.ways[i]);
        if (obj.relations) for (i = 0; i < obj.relations.length; i++) pbf.writeMessage(4, Relation.write, obj.relations[i]);
        if (obj.changesets) for (i = 0; i < obj.changesets.length; i++) pbf.writeMessage(5, ChangeSet.write, obj.changesets[i]);
    }
};

export const StringTable = {
    read(pbf: Pbf, end?: number): IStringTable {
        return pbf.readFields(StringTable._readField, {s: []}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.s.push(pbf.readBytes());
    },
    write(obj: IStringTable, pbf: Pbf): void {
        if (obj.s) for (var i = 0; i < obj.s.length; i++) pbf.writeBytesField(1, obj.s[i]);
    }
};
// ...
export const ChangeSet = {
    read(pbf: Pbf, end?: number): IChangeSet {
        return pbf.readFields(ChangeSet._readField, {id: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.id = pbf.readVarint(true);
    },
    write(obj: IChangeSet, pbf: Pbf): void {
        if (obj.id) pbf.writeVarintField(1, obj.id);
    }
};

export const Node = {
    read(pbf: Pbf, end?: number): INode {
        return pbf.readFields(Node._readField, {id: 0, keys: [], vals: [], info: null, lat: 0, lon: 0}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.id = pbf.readSVarint();
        else if (tag === 2) pbf.readPackedVarint(obj.keys);
        else if (tag === 3) pbf.readPackedVarint(obj.vals);
        else if (tag === 4) obj.info = Info.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 8) obj.lat = pbf.readSVarint();
        else if (tag === 9) obj.lon = pbf.readSVarint();
    },
    write(obj: INode, pbf: Pbf): void {
        if (obj.id) pbf.writeSVarintField(1, obj.id);
        if (obj.keys) pbf.writePackedVarint(2, obj.keys);
        if (obj.vals) pbf.writePackedVarint(3, obj.vals);
        if (obj.info) pbf.writeMessage(4, Info.write, obj.info);
        if (obj.lat) pbf.writeSVarintField(8, obj.lat);
        if (obj.lon) pbf.writeSVarintField(9, obj.lon);
    }
};

export const DenseNodes = {
    read(pbf: Pbf, end?: number): IDenseNodes {
        return pbf.readFields(DenseNodes._readField, {id: [], denseinfo: null, lat: [], lon: [], keys_vals: []}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) pbf.readPackedSVarint(obj.id);
        else if (tag === 5) obj.denseinfo = DenseInfo.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 8) pbf.readPackedSVarint(obj.lat);
        else if (tag === 9) pbf.readPackedSVarint(obj.lon);
        else if (tag === 10) pbf.readPackedVarint(obj.keys_vals, true);
    },
    write(obj: IDenseNodes, pbf: Pbf): void {
        if (obj.id) pbf.writePackedSVarint(1, obj.id);
        if (obj.denseinfo) pbf.writeMessage(5, DenseInfo.write, obj.denseinfo);
        if (obj.lat) pbf.writePackedSVarint(8, obj.lat);
        if (obj.lon) pbf.writePackedSVarint(9, obj.lon);
        if (obj.keys_vals) pbf.writePackedVarint(10, obj.keys_vals);
    }
};

export const Way = {
    read(pbf: Pbf, end?: number): IWay {
        return pbf.readFields(Way._readField, {id: 0, keys: [], vals: [], info: null, refs: [], lat: [], lon: []}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.id = pbf.readVarint(true);
        else if (tag === 2) pbf.readPackedVarint(obj.keys);
        else if (tag === 3) pbf.readPackedVarint(obj.vals);
        else if (tag === 4) obj.info = Info.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 8) pbf.readPackedSVarint(obj.refs);
        else if (tag === 9) pbf.readPackedSVarint(obj.lat);
        else if (tag === 10) pbf.readPackedSVarint(obj.lon);
    },
    write(obj: IWay, pbf: Pbf): void {
        if (obj.id) pbf.writeVarintField(1, obj.id);
        if (obj.keys) pbf.writePackedVarint(2, obj.keys);
        if (obj.vals) pbf.writePackedVarint(3, obj.vals);
        if (obj.info) pbf.writeMessage(4, Info.write, obj.info);
        if (obj.refs) pbf.writePackedSVarint(8, obj.refs);
        if (obj.lat) pbf.writePackedSVarint(9, obj.lat);
        if (obj.lon) pbf.writePackedSVarint(10, obj.lon);
    }
};

export const Relation = {

    read(pbf: Pbf, end?: number): IRelation {
        return pbf.readFields(Relation._readField, {id: 0, keys: [], vals: [], info: null, roles_sid: [], memids: [], types: []}, end);
    },
    _readField(tag: number, obj: any, pbf: Pbf): void {
        if (tag === 1) obj.id = pbf.readVarint(true);
        else if (tag === 2) pbf.readPackedVarint(obj.keys);
        else if (tag === 3) pbf.readPackedVarint(obj.vals);
        else if (tag === 4) obj.info = Info.read(pbf, pbf.readVarint() + pbf.pos);
        else if (tag === 8) pbf.readPackedVarint(obj.roles_sid, true);
        else if (tag === 9) pbf.readPackedSVarint(obj.memids);
        else if (tag === 10) pbf.readPackedVarint(obj.types);
    },
    write(obj: IRelation, pbf: Pbf): void {
        if (obj.id) pbf.writeVarintField(1, obj.id);
        if (obj.keys) pbf.writePackedVarint(2, obj.keys);
        if (obj.vals) pbf.writePackedVarint(3, obj.vals);
        if (obj.info) pbf.writeMessage(4, Info.write, obj.info);
        if (obj.roles_sid) pbf.writePackedVarint(8, obj.roles_sid);
        if (obj.memids) pbf.writePackedSVarint(9, obj.memids);
        if (obj.types) pbf.writePackedVarint(10, obj.types);
    },

    MemberType: {
    "NODE": {
        "value": 0,
        "options": {}
    },
    "WAY": {
        "value": 1,
        "options": {}
    },
    "RELATION": {
        "value": 2,
        "options": {}
    }
} as { [K in Relation_MemberType_Key]: { value: Relation_MemberType_Value, options: any } }


};